Galvo offset parameter
I have added two OpenScan parameters to point the center of the scan away from 0,0.

The values GalvoOffsetX and GalvoOffsetY are expressed in fractions of the field of view.  This means that the Zoom does effect their values.

The maximum shift right now is +/- 1 field of view.

I want to switch them to be in optical angle, but for now this lets us point.