Correctly list scannerDevices and detectorDevices
for abstraction scanner and detector in hardware config wizard.

previously in hardware config wizard all devices are listed under both
scanner and detector even they are not a scannerDevice or detectorDevice.

for example, TCSPC_SPC150 only works as a detectorDevice but not a
scannerDevice; while NIFPGA works as both scannerDevice and
detectorDevice. So ideally we should see both SPC150 and NIFPGA listed
under detector but only NIFPGA under scanner.